idp: move production to idp2001 for jpa change
idp2001 has already been updated to store u2f tokens in a jpa database
instead of in json files on disk.  This DNS change will make this the
active server while we reconfigure idp1001.

*** This will require current users to re-register there token ***

The storage format of the json file store and the jpa store have
difference.,  the later been a JWT with what looks like an encrypted
payload.  This plus the fact that we only have a handful of people
using u2f means it will be simpler for current users to re-register there token
after use

Change-Id: I3c6d437d038cdd8721f2c46488c4da3ed06acc71