Depool ulsfo to allow for safe reconfig of CR routers there
This is the first change in rolling out a config adjustment to our CR
routers which will affect the next-hop attribute that is sent in iBGP
messages between adjacent routers locally at a site.

The change will cause the routers to put their own loopback IP on routes
being announced, as opposed to now where the original next-hop learnt
on external routes is preserved.  This will help with certain edge
cases we have observed.  More info in the Phab task.

Bug: T295672
Change-Id: If97ce15a9e5d5d333a8903fbfd7eb3be59b3504d