wikimedia-dns: refactor for durum update
In the current setup, durum works by querying gdnsd for
*.check.wikimedia-dns.org. If the origin of the query is a Wikidough
host (specified by checkdoh-map), gdnsd returns the IP address for "yes"
(specified by checkdoh-addrs).

To prepare for updates to durum where we want to notify the user about
the type of service they are using, we offset the check for
yes.check.wikimedia-dns.org to Wikidough/dnsdist instead. If a user is
using Wikidough and they query for yes.check.wikimedia-dns.org, dnsdist
will return the IP for it depending on if it was received on the DoH
port or the DoT port.

If a user is not using Wikidough, they will then hit gdnsd directly, and
that will give them the "no" IP for durum.

Essentially what this commit does is that it tells Wikidough to take
care of yes{doh,dot}.check.wikimedia-dns.org and only leaves gdnsd to
worry about the case where people are not using Wikidough.

This not only helps improve the code complexity but also helps us
achieve the desired durum update where we want to notify the user of the
type of protocol they are using over Wikidough (DoH or DoT).

Change-Id: I317c982e3ec242d0fc11ad6d8015bc65571cfc8e