geodns: remove geo-maps-esams-offline hack
This hack redistributed much of the normal eqiad load towards
codfw and ulsfo when esams went offline, because otherwise eqiad
would be overwhelmed by the failover traffic from esams.  We now
have the capability to fail over traffic to drmrs in this case
(coming in the next patch), so we can remove this awful exception
case.

Bug: T304089
Change-Id: I7585056c45937d45c2d284bd73315619fee9d033