geodns: add drmrs fallback for esams to whole map
Add drmrs as the next entry behind esams in all of the map's
datacenter lists (except for those already on drmrs primary: 3
internal networks + CY, PT, ES, FR).  No new sets of users are
mapped to drmrs as their primary option in this patch.

This replaces the geo-maps-esams-offline hack (removed in previous
commit) with the ability to shift esams traffic to drmrs in the
case that esams needs to be depooled, which is an important
resiliency win while we work on the mapping for expanding drmrs
use as a primary target for users.

Bug: T304089
Change-Id: Id8625e34c84e1e94df4779021351427ee2e26ffe