switch policy.wikimedia.org back from Wordpress to WMF DNS
This site once was hosted in-house, then moved to external hosting
at Wordpress VIP and is now requested to come back to WMF again.

The new request to turn it into redirects is T310738 but there is
also longstanding ticket T132104 which this would solve as well.

This should go after  I8dc6d27dccfcb5e6 which adds the rewrite rules
to apache config on the appserver cluster.

Also will follow-up with httpbb tests.

Bug: T310738
Bug: T132104
Change-Id: I7e56bc227f8162f1e02882a4c26ea232cfe1bca0