add gerrit-replica-new.wikimedia.org, point to 208.80.153.104
As part of the "decommission mission" sprint with releng
we are replacing old hardware gerrit2001.

This currently hosts gerrit-replica.wikimedia.org.

The replacement host gerrit2002 is waiting to take over.

In order to replicate data correctly we would like to tell
the main production gerrit server to start replicating (additionally)
to the new up-and-coming replica.

gerrit2001 has 2 IP addresses configured, 208.80.153.106 (gerrit2001)
and 208.80.153.107 (gerrit-replica).

So to copy that we need to add a second IP to gerrit2002. This will
temporarily be called gerrit-replica-new.

Once the migration is over gerrit-replica-new will switch to
gerrit-replica.

This is exactly like it was done for gitlab as well where we temp
had replica-new and replica-old even.

We have been assigned 208.80.153.104 / 2620:0:860:4:208:80:153:104
at T313972#8115732.

Bug: T313250
Bug: T243027
Bug: T313972
Change-Id: Id41c8c245fc7224217ce75fe29e63370df4e4979