gitlab: add reserved service IP 208.80.153.8, point to replica-new
Arnold has worked with Riccardo to get a new service IP reserved
in netbox.

Riccardo assigned 208.80.153.8 and the mapped IPv6 equivalent.

https://netbox.wikimedia.org/ipam/ip-addresses/11647/

In Netbox it has the comment "Keep manual DNS used for GitLab SSH
service on gitlab2002" and the entry is for 208.80.153.8/27 and
not as a VIP.

I do want to point out I need the same thing for gerrit-replica
but in netbox that is added/reserved in a different way, as a VIP
even though it isn't one. So I can either follow the example
from gerrit or from gitlab but the point is they should all be the
same. There should not be different ways they are in netbox.

Bug: T307142
Change-Id: If695b0958fb756f7ec58be27eb5a3363f5161c58