delete varnish service alias for phab2001-aphlict
We have a hardcoded host name inside a Varnish service name here.

And phab2001 is about to be removed.

We would either just keep the "2001" in the varnish service name
but point it over to phab2002, right with I2801ff0761516.

Or we could say this hostname in the service name is horrible
and get into the part how to rename that.

Right now I am leaving it here as is to discuss and remind me.

edit: after talking with bblack on IRC now just removing this line

Bug: T322250
Change-Id: I2fca478c65d4238845dfb39ce58c2e873cb4b213