remove phab1001-aphlict.eqiad.wmnet
I don't think this is used.

This is an alias for phab1001 but aphlict does not run there anymore
since it moved to its own dedicated VM, aphlict1001.

In I2fca478c65d4 the equivalent in codfw was deleted
and it had comments like it had to stay around for
varnish puppetization woes in the past.

bblack did some research on that one and found that
it didn't seem to be used by anything.

When aphlict was moved out of Phabricator servers into
its own VM, aphlict1001.eqiad.wmnet, it also got a
discovery DNS record.

So nowadays aphlict.discovery.wmnet is a thing and used and it
currently points to aphlict1001.eqiad.wmnet.

aphlict2001.codfw.wmnet does not exist yet, but should (T322369).

But this phab1001-aphlict.eqiad.wmnet I think can be deleted
as well.

Bug: T280597
Bug: T322369
Change-Id: Id507a0883ba8b2a0fdecf7ede124a772cab017f1