Add spf record for gitlab.wikimedia.org
This attempts to address a report that a legitimate mail from our gitlab
instance ended up in a user's spam folder. The first attempt to fix this
should be adding SPF records.

For this, we just include the existing wikimedia.org SPF record:

```
"v=spf1 ip4:91.198.174.0/24 ip4:208.80.152.0/22 ip6:2620:0:860::/46 include:_spf.google.com ip4:74.121.51.111 ~all"
```

I'm not sure if this is the best approach here, it may be better to add
only the IPs from the gitlab instances, but that becomes an ugly record
to look at and difficult to maintain if we change our gitlab hosts.

Bug: T328642
Change-Id: I67c28206928b450127d070859857ddb14d9d522c