switch planet from eqiad to codfw
This switches the planet.discovery.wmnet record
from the eqiad to the codfw instance.

We don't have geo DNS/active active but we do
have a discovery record.

The service runs updates and pulls data on
both instances at all times and the webserver runs on both.

In ATS traffic layer we already use the discovery name
so there are no hardcoded instance names, also not in the
certificate SAN.

No change in puppet is needed, only DNS.

When should I make the switch? Anytime or coordinated?

How would this be added to cookbooks though I don't know.

Bug: T330091
Change-Id: I3f172f8e79b4fef88166ec47732369947aada940