switch peopleweb from eqiad to codfw
This changes the peopleweb.discoveryw.wmnet name from
the eqiad to the codfw backend.

Similar to planet this is also a service that has a
discovery name but no geo DNS.

We are already avoiding hardcoded host names but
here one server is clearly the active one as in
"the source of rsyncing data" and "where people should
upload files".

Other than that it's just a webserver.

So this requires one more puppet change to switch the
direction of the data sync but that is all.

In addition to peopleweb.discovery.wmnet, which is used
in ATS config, there is a "vanity name" for end users
just called "people".

But there was only people.eqiad.wmnet and people.codfw.wmnet
did not exist. So adding that just for consistency. But probably
nobody expected this to be a thing yet.

Finally, people2001 has meanwhile been replaced by people2002
but we failed to update the commented out line here, so that's
what that changes as well.

Bug: T330091
Change-Id: I073e70e18ad7279593b11899cbbd2986a959f0db