re-introduce webserver-misc-static.discovery.wmnet
In the past we had separate pairs of VMs, one for
static sites and one for applications.

Later these were merged into a single pair of machines.

Now we want the ability back to switch only the static
sites between data centers but not touch the apps.

Therefore we re-introduce the 2 separate names to have that
flexibility even though it does NOT mean we add new VMs again.

The switch from eqiad to codfw is already included so after
that we can simply switch individual sites from -apps to -static
and it will mean they are served from codfw.

Bug: T330090
Change-Id: Ied8a689f30283b95fa9718d728359cc2ae3203bc