Add forward and reverse entries for aqs.svc.codfw.wmnet
This is part of the effort to make AQS a multi-dc service. This change
goes along with a correspnding change to the service catalog in
https://gerrit.wikimedia.org/r/c/operations/puppet/+/894017

Note that there are still outstanding questions about whether we
are ready to pool AQS in codfw, but adding this DNS record is only an
intermediate step.

The linked ticket explains the outstanding questions.

Bug: T331115
Change-Id: I9c9a0842c1c29065841a6eea05fe45893a85e1c2