remove miscweb2002, was commented out fail-over machine
This was commented out as it's not an active server
but still make it clear this does not exist anymore
and it's not an option to fail-over to it.

The actual replacement would be miscweb2003 but also
there is only a single site left on miscweb1002, the
other side of the miscweb/buster pair,which is iegreview.

We are currently determining whether iegreview will be shut down
or not.. if it is we just delete miscweb1002, if it's not we
migrate it to miscweb2003.

Neither case we use miscweb2002 anymore.

Bug: T334024
Change-Id: Ic2e2375f60dac93cbeb4d59643408602ec2c9adf