gerrit: switch service IP, turn new into current and current into old
For the day of the actual switch of gerrit.wikimedia.org from old to
new hardware, which involves a new service IP.

Turn "gerrit-new" into "gerrit" and "gerrit" into "gerrit-old".

That way, as discussed on IRC, there is still a way to access the old server
for a little bit.

In Netbox these are placeholders for manual management so I think it's fine
but we should just update the comments there.

And then soon we will give the old IP back into the pool.

Bug: T326368
Change-Id: Id13e858148bcd4f0e5716ff72872edcffd136efa