Convert kubestagemaster from CNAME to A record (#7)
What?
We have assigned VIPs for kubestagemaster as
it is now an lvs service.

Why?
kubestagemaster used to be a CNAME when we had
one master in this cluster, but now we will be
having two, thus we need them behind an lvs service.

Related: I4088263d3119addb7ff483d684069f8560315d16

Bug: T329827
Change-Id: I6f633ffce5df8c6909ca25357c262e6787301d4b