Remove IPv6 reverse include statements for old mr1-codfw CR links
mr1-codfw connection to wider network has been moved from old xlink
vlan + ospf setup to BGP to lsw1-a2-codfw.  This removes the include
entries for the netbox-generated reverse records for the old IPs.

Also removed the list of all link subnets from the IPv4 file
153.80.208.in-addr.arpa.  This had the links I removed but I think
it is too cumbersome to maintain a list there, so I removed them
all.  Info is in Netbox now and easy to display.

Bug: T348164
Change-Id: I372300bbbc86dbc28021f0271908b8cb281d2721