ores: Remove old ORES DNS entries
Why:

These aren't working anymore. There is no ORES service, nor anything in
puppet.git that references this IP. There is a reference for changeprop
in deployment-charts but it's just a dangling netpol. There is also a
codepath that obviously now fails since nothing listens on 10.2.2.10,
the /sys/ores in changeprop's configuration that will be removed too.

What:

Just remove them

Change-Id: I2bf9a344e080e023295277942db50ba184843e26