create wikipedia-pl-sysop.wikimedia.org
The original request was to create sysop-pl.wikipedia.org
but this was controversial as we shouldn't add special projects
like this into the wikipedia.org domain.

Therefore the second best seemed to be this (because there
can be special sysop wikis for other projects, not just Wikipedia).

Updated after review comment to match standard for mailing lists
like in https://meta.wikimedia.org/wiki/Mailing_lists/Standardization.

Bug: T361041
Change-Id: Ie2f85841f63d68ecd7fe4795d2cb622595888e2a