geo-maps: add magru to geo maps
This commit adds the new DC magru in Brazil. Since we don't plan on
turning this on yet and we don't have latency data for magru, we start
by adding it at the end of the preferred DC list for all regions.

When we turn on the site, we will add a regional override with magru as
the preferred DC and then slowly update the rest of the geo-map.

Bug: T346722
Change-Id: I93fe45bed44583c86680b5595c481181e048282b