Add reverse DNS entries for new codfw private subnets
IPs with dns names in these subnets have been assigned in Netbox
so we need to make sure the correct includes as in the zone
files to allow the netbox dns name generation to keep working.

I also moved one or two blocks to keep the numbering in order,
and removed the comment pointing to Netbox wikitech under
every single one as I think we all agree that's too verbose.

Bug: T364095
Change-Id: I8c1850893aa183277a26b9e506f405e5d49d818c