move linkrecommendation service IP in place, fix outdated comments
The service IP for linkrecommendation was not added in the right
placed, sorted numerically.

Remaining comments that this used to be the eventbus IP made
it look like this is free for use and you would only notice
it's actually already used once looking at the reverse records.

Clean that up.

Change-Id: I9b626fad2cacf1096e2dfe85d04b29372ea9ef58