geo-maps: define initial mapping for South America (magru)
This commit sets the new Brazil DC magru to be the default site for
traffic from Bolivia and Paraguay, to slowly and progressively ramp up
traffic to the new DC. The two primary goals for doing this with these
two countries/regions is to test connectivity to the new DC with live
traffic and to warm the caches before we start sending more traffic to
it. These two countries also form the first of three cohorts in which we
will slowly ramp up traffic to the new DC.

For more discussion on why we picked these two regions, please see the
task below.

Bug: T346722
Bug: T359054
Change-Id: Ia5c25bfded1f429c345fe2b8e5497f05ca3bea30