rename gitlab-replica-old to gitlab-replica-b
We have discussed that the name "old" isn't a good fit for what this is.

It's not actually old, it's just one of 2 replicas that should be equal.

So the idea is to call them "replica A" and "replica B" to make that
clear and not confuse anyone thinking this is somehow an older gitlab
or OS version or to be decom'ed soon.

Change-Id: I0762b82ffe810e44507805defdada904cecee583