delete cache.wikimedia.org
This has been flagged as an "unused subdomain".

Searching in codesearch shows nothing that uses it.

Neither does grepping the puppet repo.

And when I looked at git history I had to go back many many commits
and ended up at a time before DNS was even in version control.

The suggestion is to just drop it.

But given the nature of it and how scary this sounds it will need
double checking by traffic team.

Bug: T367012
Change-Id: Ie8eaaa8cdc053bfd3edac8e29ffa57b29f0e4da0