dns: provision airflow-analytics-test domain
We are in the process of moving Airflow services to the dse-k8s cluster and we shall have all the
instances on their own wikimedia domain pointing to the dse-k8s cluster.
The analytics-test instance will be the first to migrate so we are setting up the domain first and
the service record.

Bug: T371209
Bug: T371210
Change-Id: I27ae86c8539cef0cc73bd3cf7e908321249af54c