Delegate Kubernetes POD IP reverse ranges to k8s control-plane nodes
As part of the I/F Foundations hackathon we have a group working to
improve our debugging ability when we have issues with DB hosts.

One element of this is being able to better identify what kubernetes
PODs are using the IP addresses we see in logs.  To expose this data
in a reasonable way cdanis has been working to get coredns running
on the control-plane hosts for each kubernetes cluster.

This patch delegates the reverse zones corresponding to IP address
space used by Kubernetes PODs to those hosts.  It is added here
manually so we can continue the overall work, however we will
continue to work with stakeholders to get a process to generate
and add this data to the DNS hosts automatically.

Bug: T376291
Change-Id: I696c445f96143e5989eec3448fe1e2bb2b139780