Raise the weight of all analytics mariadb replica srv records
We are going to be making more use of these SRV records in order to
allow mediawiki to look up the correct host and port for running dumps.

While the records are correct at the moment, having a weight of 1 does
not allow for a lower weight server to be added in. Therefore, we
increase all weights to 100.

Bug: T382947
Change-Id: I38ea3b1eaa52f0ec610fe465705383186cc8062c