add ingress service aliases for codesearch on k8s-aux
Following the setup of jaeger and the docs on Wikitech,
set svc aliases to point to the k8s ingress, on the aux cluster.

Note that the entire section for ingress aliases in codfw
on aux was missing and I am adding it. Shouldn't jaeger
also have these?

Namespace with TLS cert names added in I5f318eee18403cb.

Bug: T268199
Change-Id: Ic88368ebefa2c9d0021da34b10dc3317811d0e85