add k8s ingress service aliases for jaeger in codfw
When adding a new service (codesearch) in the aux cluster
and following docs I noticed that the jaeger service only
has ingress service aliase in eqiad but not in codfw.

An entire section for aliases in codfw on the aux cluster
seemed to be missing.

Assuming these records should be added for jaeger.

Bug: T345894
Bug: T381417
Change-Id: I1edc541e85220e9a769caf7d868111a1dceb1bd7