utils: add a script to generate HTTPS TYPE65 records for ECH
gdnsd natively does not support HTTPS resource records (TYPE65). But it
does support specifying the TYPE and a DNS record in wire format, so we
generate those with the script for ECH and then paste them in the
relevant zone file.

There is no direct integration with gdnsd yet and that's fine. We really
don't need it for the purpose of this experiment.

Bug: T205378
Change-Id: I5964db8fa99815fc2107e97ebd1ef599d02399bd