wikimedia-ech.org: update zone file and add A/AAAA records
Per the RFC [0], outer ECH should also provide a valid TLS cert and it
currently does not since we point this domain to ncredir. Given that
there is only one service behind this, update the zone file to point
wikimedia-ech.org to the check service so that a valid cert is returned.

The check service is already set up with the correct SNI and listening
on the right IP addresses.

[0] - https://datatracker.ietf.org/doc/draft-ietf-tls-esni/

Bug: T205378
Change-Id: Ia6453719f696a337044a71d03af3122118f6c8a3