wikimedia-dns.org: bump up TTL for TYPE65 record (60 to 600)
Since this service is in production, there is no reason to have a short
TTL so use a proper 600 second TTL instead.

This may also help with debugging of certain connection issues in which
the TTL is short so the lookup for HTTPS records fails and ECH is not
attempted.

Change-Id: I3ec762764c22699f40c8eda2f33fbb73c8b01954