Revert^2 "search: cname specific search clusters to the lvs pool"
There are three search clusters, but up until now they have
always been accessed through different ports on a single
name.
This series of patches adds named services for the other
two clusters so we can address them individually.

This reverts commit 5d9c1f83772b8ebac07f2edc3477c4de43ab01da. It
was initially merged with a second patch that had deployment
errors and was reverted, but should be reasonable to ship.

https://wikitech.wikimedia.org/wiki/LVS#DNS_changes_(svc_zone_only)

Depends-On: Ieefc6546d8f3cdb7b114d69882d0f5021ca9aca4
Change-Id: Ife42e03f458a20474391a4d6e2bdb54678e5ed7d