Add tj to langlist helper
'tj' is NOT a valid ISO-639 language code.

But this is what is considered a good thing in this special case
because it means it will not conflict with another language
in the future.

The intention is to redirect users from "tj" to "tg" in I8f1d515e562d965b5e1f.

tj is the country code for Tajikistan while tg is the language
code for Tajik; therefore they often get mixed up.

The request is to redirect one to the other for convenience.

Bug: T393803
Change-Id: I5b4913d90bf4719d69d3c1dabe36574de5538b76