wmnet: direct swift-ro DYNA record to failoid
The swift-r[ow] discovery services have repeatedly been a source of
confusion.

Although both services are ostensibly unused, we would like to begin the
turndown process with a "soft" (reversible) shutdown.

For swift-rw, an active-passive service, this has already started - the
service is depooled everywhere, and thus resolves to failoid.

For swift-ro, we would like to use the same trick as used previously for
appservers-ro [0] - since failoid is a geoip resource like any other, we
can achieve the same by overriding the swift-ro DYNA record.

[0] I76a3e846769b6cb6986bf7c31a99f29066c928cc

Bug: T376237
Change-Id: Iaa3f40d049aac58400830bee7157f0bbe1796513