DNS: Add druid-public-coordinator record
We are adding the druid coordinator service to lLVs as part of our ongoing work to use single urls
per service as opposed to using hardcoded hosts names which cause issues during maintenance
downtimes and general host issues.

This patch adds the druid-public-coordinator to service list.
The ip 10.2.2.15 is currently in a reserved state and shall be set to active as we proceed.
This also reserves the ips for codfw

Bug: T406222
Change-Id: I8a8cee286745b71604c797170a8b34324117b706