add discovery records for gerrit as CNAMEs to public names
Per discussion in today's meeting re: moving Gerrit behind the CDN.

We will need discovery.wmnet records but gerrit servers only have
2 public IPs.

We can still create CNAMEs to point to them though.

So using the section that is normally for services without geoip
but multiple backends.

One for gerrit and one for gerrit-replica.

Should not have any effect if this exists before the rest of the setup
so we should be able to get this out of the way early.

Bug: T365259
Change-Id: Iea0cf8d0cf5a44f3aaf3d73f716793cf928bbe8d